Searching in Tab Tray working with empty screen state
