+++ Client/Frontend/Browser/TabManager.swift
@@ -178,11 +178,11 @@ class TabManager: NSObject {
         
         let allTabs = allTabs.filter { $0.type == type }
         
-        if let query = query {
-            return allTabs.filter { $0.type == type &&
-                (($0.lastTitle?.contains(query) != nil) || (($0.canonicalURL?.absoluteString.contains(query)) != nil)) }
+        if let query = query, !query.isEmpty {
+            // Last title is the only data that will be present on every situation
+            return allTabs.filter { $0.lastTitle?.contains(query) ?? false }
         } else {
-            return allTabs.filter { $0.type == type }
+            return allTabs
         }
     }
